requirements.txt
streamlit
ultralytics
opencv-python
scipy
transformers
pydub
scipy
audioread
requests
tqdm
st-annotated-text
tensorflow
tf-keras
streamlit_option_menu